+++ src/main/java/pl/asie/computronics/ClientProxy.java
@@ -63,12 +63,10 @@ public class ClientProxy extends CommonProxy {
 			y = p.readDouble(),
 			z = p.readDouble();
 		float force = p.readFloat();
+		boolean destroyBlocks = p.readByte() != 0;
 		Minecraft minecraft = Minecraft.getMinecraft();
 		SelfDestruct explosion = new SelfDestruct(minecraft.theWorld,
-			null, x,
-			y,
-			z,
-			force);
+			null, x, y, z, force, destroyBlocks);
 		int size = p.readInt();
 		ArrayList<BlockPos> list = new ArrayList<BlockPos>(size);
 		int i = (int) x;
@@ -104,5 +102,10 @@ public class ClientProxy extends CommonProxy {
 			IntegrationOpenComputers.upgradeRenderer = new UpgradeRenderer();
 		}
 		MinecraftForge.EVENT_BUS.register(IntegrationOpenComputers.upgradeRenderer);
+
+		if(IntegrationOpenComputers.mountableRenderer == null) {
+			IntegrationOpenComputers.mountableRenderer = new RackMountableRenderer();
+		}
+		MinecraftForge.EVENT_BUS.register(IntegrationOpenComputers.mountableRenderer);
 	}
 }