+++ src/main/java/pl/asie/computronics/integration/tis3d/module/ModuleBoom.java
@@ -120,6 +120,7 @@ public class ModuleBoom extends ComputronicsModule {
 
 			FontRendererAPI.drawString(s);
 		}
+		GL11.glDisable(GL11.GL_BLEND);
 	}
 
 	public static class BoomHandler {
@@ -134,7 +135,7 @@ public class ModuleBoom extends ComputronicsModule {
 			}
 			for(Casing casing : boomQueue) {
 				if(casing != null && casing.getCasingWorld() != null) {
-					SelfDestruct.goBoom(casing.getCasingWorld(), casing.getPosition());
+					SelfDestruct.goBoom(casing.getCasingWorld(), casing.getPosition(), true);
 				}
 			}
 			boomQueue.clear();