+++ src/main/java/pl/asie/computronics/oc/ColorfulUpgradeHandler.java
@@ -13,6 +13,7 @@ import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import org.lwjgl.opengl.GL11;
 import pl.asie.computronics.item.ItemOpenComputers;
+import pl.asie.computronics.oc.driver.RobotUpgradeColorful;
 import pl.asie.computronics.reference.Mods;
 
 import java.util.Locale;