+++ src/main/java/pl/asie/computronics/oc/driver/DriverCardBoom.java
@@ -1,15 +1,14 @@
-package pl.asie.computronics.oc;
+package pl.asie.computronics.oc.driver;
 
 import li.cil.oc.api.Network;
-import li.cil.oc.api.driver.EnvironmentHost;
 import li.cil.oc.api.machine.Arguments;
 import li.cil.oc.api.machine.Callback;
 import li.cil.oc.api.machine.Context;
 import li.cil.oc.api.network.Component;
+import li.cil.oc.api.network.EnvironmentHost;
 import li.cil.oc.api.network.Message;
 import li.cil.oc.api.network.Node;
 import li.cil.oc.api.network.Visibility;
-import li.cil.oc.api.prefab.ManagedEnvironment;
 import net.minecraft.nbt.NBTTagCompound;
 import pl.asie.computronics.Computronics;
 import pl.asie.computronics.reference.Mods;
@@ -18,21 +17,25 @@ import pl.asie.computronics.util.boom.SelfDestruct;
 /**
  * @author Vexatos
  */
-public class DriverCardBoom extends ManagedEnvironment {
+public class DriverCardBoom extends ManagedEnvironmentWithComponentConnector {
 
 	protected final EnvironmentHost container;
 
 	public DriverCardBoom(EnvironmentHost container) {
 		this.container = container;
-		this.setNode(Network.newNode(this, Visibility.Neighbors).
-			withComponent("self_destruct").
-			create());
+		createNode();
 		if(this.node() != null) {
 			initOCFilesystem();
 		}
 	}
 
-	private li.cil.oc.api.network.ManagedEnvironment oc_fs;
+	protected void createNode() {
+		this.setNode(Network.newNode(this, Visibility.Neighbors).
+			withComponent("self_destruct").
+			create());
+	}
+
+	protected li.cil.oc.api.network.ManagedEnvironment oc_fs;
 
 	private void initOCFilesystem() {
 		oc_fs = li.cil.oc.api.FileSystem.asManagedEnvironment(li.cil.oc.api.FileSystem.fromClass(Computronics.class, Mods.Computronics, "lua/component/self_destruct"),
@@ -54,7 +57,7 @@ public class DriverCardBoom extends ManagedEnvironment {
 			// computer.
 			node.disconnect(oc_fs.node());
 		} else if(node == this.node()) {
-			this.time = -1;
+			setTime(-1);
 			// Remove the file system if we are disconnected, because in that
 			// case this method is only called once.
 			oc_fs.node().remove();
@@ -67,12 +70,16 @@ public class DriverCardBoom extends ManagedEnvironment {
 		if((message.name().equals("computer.stopped")
 			|| message.name().equals("computer.started"))
 			&& node().isNeighborOf(message.source())) {
-			this.time = -1;
+			setTime(-1);
 		}
 	}
 	// Boom code
 
-	private int time = -1;
+	protected int time = -1;
+
+	protected void setTime(int time) {
+		this.time = time;
+	}
 
 	@Callback(doc = "function([time:number]):number; Starts the countdown; Will be ticking down until the time is reached. 5 seconds by default. Returns the time set")
 	public Object[] start(Context context, Arguments args) {
@@ -83,7 +90,7 @@ public class DriverCardBoom extends ManagedEnvironment {
 		if(fuse > 100000) {
 			throw new IllegalArgumentException("time may not be greater than 100000");
 		}
-		this.time = (int) Math.round(Math.floor(fuse * 20));
+		setTime((int) Math.round(Math.floor(fuse * 20)));
 		return new Object[] { fuse };
 	}
 
@@ -107,7 +114,7 @@ public class DriverCardBoom extends ManagedEnvironment {
 			oc_fs.node().load(nbt.getCompoundTag("oc:fs"));
 		}
 		if(nbt.getBoolean("ticking")) {
-			this.time = nbt.getInteger("time");
+			setTime(nbt.getInteger("time"));
 		}
 	}
 
@@ -133,11 +140,15 @@ public class DriverCardBoom extends ManagedEnvironment {
 		if(this.time < 0) {
 			return;
 		}
-		this.time--;
 		if(this.time <= 0) {
 			//Bye bye.
-			this.time = -1;
-			SelfDestruct.goBoom(container.world(), container.xPosition(), container.yPosition(), container.zPosition());
+			goBoom();
+		} else {
+			--this.time;
 		}
+	}
+
+	protected void goBoom() {
+		SelfDestruct.goBoom(container.world(), container.xPosition(), container.yPosition(), container.zPosition(), true);
 	}
 }